Split error types from lib.rs
